move emptyStatePage component to common folder, use doc link from constant
