+++ app/src/pages/inside/common/emptyStatePage/emptyStatePage.scss
