+++ app/src/pages/inside/common/emptyStatePage/index.js
