+++ app/src/pages/inside/projectSettingsPageContainer/content/patternAnalysis/patternAnalysis.jsx
@@ -16,20 +16,22 @@
 
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { addPatternAction, patternsSelector } from 'controllers/project';
 import { useTracking } from 'react-tracking';
-import { getSaveNewPatternEvent, SETTINGS_PAGE_EVENTS } from 'components/main/analytics/events';
-import { PROJECT_SETTINGS_PATTERN_ANALYSIS_EVENTS } from 'analyticsEvents/projectSettingsPageEvents';
-import { hideModalAction, showModalAction } from 'controllers/modal';
-import { STRING_PATTERN } from 'common/constants/patternTypes';
 import { useIntl } from 'react-intl';
-import { canUpdateSettings } from 'common/utils/permissions';
-import { userRolesSelector } from 'controllers/pages';
 import PropTypes from 'prop-types';
+
+import { EmptyStatePage } from 'pages/inside/common/emptyStatePage';
+import { addPatternAction, patternsSelector } from 'controllers/project';
+import { getSaveNewPatternEvent, SETTINGS_PAGE_EVENTS } from 'components/main/analytics/events';
+import { hideModalAction, showModalAction } from 'controllers/modal';
+import { userRolesSelector } from 'controllers/pages';
+import { PROJECT_SETTINGS_PATTERN_ANALYSIS_EVENTS } from 'analyticsEvents/projectSettingsPageEvents';
+import { STRING_PATTERN } from 'common/constants/patternTypes';
+import { canUpdateSettings } from 'common/utils/permissions';
 import { docsReferences } from 'common/utils';
+
 import { SettingsPageContent } from '../settingsPageContent';
 import { PatternAnalysisContent } from './patternAnalysisContent';
-import { EmptyStatePage } from '../emptyStatePage/';
 import { messages } from './messages';
 
 export const PatternAnalysis = ({ setHeaderTitleNode }) => {
